requirements.txt
google-generativeai==0.3.2
requests==2.31.0
PyYAML==6.0.1